+++ tests/test_sdr.py
@@ -1,14 +1,12 @@
 """InVEST SDR model tests."""
-import unittest
-import tempfile
-import shutil
 import os
+import shutil
+import tempfile
+import unittest
 
-import pygeoprocessing
 import numpy
-from osgeo import ogr
-from osgeo import osr
-from osgeo import gdal
+import pygeoprocessing
+from osgeo import gdal, ogr, osr
 
 REGRESSION_DATA = os.path.join(
     os.path.dirname(__file__), '..', 'data', 'invest-test-data', 'sdr')
@@ -221,6 +219,14 @@ class SDRTests(unittest.TestCase):
             args['workspace_dir'], 'watershed_results_sdr.shp')
         assert_expected_results_in_vector(expected_results, vector_path)
 
+        # We only need to test that the drainage mask exists.  Functionality
+        # for that raster is tested elsewhere
+        self.assertTrue(
+            os.path.exists(
+                os.path.join(
+                    args['workspace_dir'], 'intermediate_outputs',
+                    'what_drains_to_stream.tif')))
+
     def test_regression_with_undefined_nodata(self):
         """SDR base regression test with undefined nodata values.
 
@@ -346,8 +352,8 @@ class SDRTests(unittest.TestCase):
 
     def test_missing_lulc_value(self):
         """SDR test for ValueError when LULC value not found in table."""
-        from natcap.invest.sdr import sdr
         import pandas
+        from natcap.invest.sdr import sdr
 
         # use predefined directory so test can clean up files during teardown
         args = SDRTests.generate_base_args(self.workspace_dir)
@@ -368,6 +374,48 @@ class SDRTests(unittest.TestCase):
             "The missing values found in the LULC raster but not the table"
             " are: [2.]" in str(context.exception),
             f'error does not match {str(context.exception)}')
+
+    def test_what_drains_to_stream(self):
+        """SDR test for what pixels drain to a stream."""
+        from natcap.invest.sdr import sdr
+
+        srs = osr.SpatialReference()
+        srs.ImportFromEPSG(26910)  # NAD83 / UTM zone 11N
+        srs_wkt = srs.ExportToWkt()
+        origin = (463250, 4929700)
+        pixel_size = (30, -30)
+
+        flow_dir_mfd = numpy.array([
+            [0, 1],
+            [1, 1]], dtype=numpy.float64)
+        flow_dir_mfd_nodata = 0  # Matches pygeoprocessing output
+        flow_dir_mfd_path = os.path.join(self.workspace_dir, 'flow_dir.tif')
+        pygeoprocessing.numpy_array_to_raster(
+            flow_dir_mfd, flow_dir_mfd_nodata, pixel_size, origin, srs_wkt,
+            flow_dir_mfd_path)
+
+        dist_to_channel = numpy.array([
+            [10, 5],
+            [-1, 6]], dtype=numpy.float64)
+        dist_to_channel_nodata = -1  # Matches pygeoprocessing output
+        dist_to_channel_path = os.path.join(
+            self.workspace_dir, 'dist_to_channel.tif')
+        pygeoprocessing.numpy_array_to_raster(
+            dist_to_channel, dist_to_channel_nodata, pixel_size, origin,
+            srs_wkt, dist_to_channel_path)
+
+        target_what_drains_path = os.path.join(
+            self.workspace_dir, 'what_drains.tif')
+        sdr._calculate_what_drains_to_stream(
+            flow_dir_mfd_path, dist_to_channel_path, target_what_drains_path)
+
+        # 255 is the byte nodata value assigned
+        expected_drainage = numpy.array([
+            [255, 1],
+            [0, 1]], dtype=numpy.uint8)
+        what_drains = pygeoprocessing.raster_to_numpy_array(
+            target_what_drains_path)
+        numpy.testing.assert_allclose(what_drains, expected_drainage)
 
     @staticmethod
     def _assert_regression_results_equal(